refactor: moved track story method to StoriesManager
